Implemented bands API and subscriptions API

See merge request igeneralov/anton-bulgakov-internship!10

+++ src/main/java/com/dataart/playme/controller/rest/PostController.java
@@ -4,6 +4,7 @@ import com.dataart.playme.controller.binding.annotation.CurrentMusician;
 import com.dataart.playme.dto.CreateCommentDto;
 import com.dataart.playme.dto.PostRequestDto;
 import com.dataart.playme.dto.PostResponseDto;
+import com.dataart.playme.exception.ConflictException;
 import com.dataart.playme.model.Band;
 import com.dataart.playme.model.Comment;
 import com.dataart.playme.model.Musician;
@@ -57,7 +58,10 @@ public class PostController {
                               @Valid @RequestBody CreateCommentDto dto,
                               BindingResult bindingResult) {
         if (!bindingResult.hasErrors()) {
-            return postService.createComment(post, musician, dto);
+            if (bandService.isActiveBand(post.getBand())) {
+                return postService.createComment(post, musician, dto);
+            }
+            throw new ConflictException("Band is disabled");
         }
         throw new BadRequestException("Invalid data");
     }
@@ -65,7 +69,10 @@ public class PostController {
     @DeleteMapping("/{post}")
     public Post deletePost(@PathVariable Post post,
                            @CurrentMusician Musician musician) {
-        return postService.deletePost(post, musician);
+        if (bandService.isActiveBand(post.getBand())) {
+            return postService.deletePost(post, musician);
+        }
+        throw new ConflictException("Band is disabled");
     }
 
     @DeleteMapping("/{post}/comments/{comment}")
